Маппинг объявлений создан (без байтов изображения)

+++ example-for-graduate-work-main/src/main/java/ru/skypro/homework/repository/UserRepository.java
@@ -6,8 +6,12 @@ import org.springframework.stereotype.Repository;
 import ru.skypro.homework.model.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     @Query(value = "SELECT u FROM User u WHERE u.email = :userName")
     public User findUserByUserName(String userName);
+
+
+
+
 }
